refactor: extract file path completion microcrate
