+++ Cargo.toml
@@ -53,6 +53,7 @@ members = [
     "crates/perl-lsp-workspace-symbols",
     "crates/perl-lsp-completion",
     "crates/perl-lsp-completion-item",
+    "crates/perl-lsp-file-path-completion",
     "crates/perl-lsp-code-actions",
     "crates/perl-lsp-folding",
     "crates/perl-lsp-ast-utils",
@@ -181,6 +182,7 @@ allow = [
     "perl-lsp-folding",
     "perl-lsp-completion",
     "perl-lsp-completion-item",
+    "perl-lsp-file-path-completion",
     "perl-lsp-inlay-hints",
 
     # Tier 3 — Module resolution chain
@@ -379,6 +381,7 @@ perl-lsp-workspace-symbols = { path = "crates/perl-lsp-workspace-symbols", versi
 perl-lsp-rename = { path = "crates/perl-lsp-rename", version = "0.11.0" }
 perl-lsp-completion = { path = "crates/perl-lsp-completion", version = "0.11.0" }
 perl-lsp-completion-item = { path = "crates/perl-lsp-completion-item", version = "0.11.0" }
+perl-lsp-file-path-completion = { path = "crates/perl-lsp-file-path-completion", version = "0.11.0" }
 perl-lsp-code-actions = { path = "crates/perl-lsp-code-actions", version = "0.11.0" }
 perl-lsp-folding = { path = "crates/perl-lsp-folding", version = "0.11.0" }
 perl-diagnostics-codes = { path = "crates/perl-diagnostics-codes", version = "0.11.0" }
