+++ crates/perl-lsp-file-path-completion/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "perl-lsp-file-path-completion"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license = "MIT OR Apache-2.0"
+repository.workspace = true
+homepage.workspace = true
+description = "SRP microcrate for secure Perl LSP file path completion"
+readme = "README.md"
+documentation = "https://docs.rs/perl-lsp-file-path-completion"
+keywords = ["perl", "lsp", "completion", "paths", "ide"]
+categories = ["development-tools", "text-editors"]
+include = [
+  "src/**",
+  "Cargo.toml",
+  "LICENSE*",
+  "README.md"
+]
+
+[dependencies]
+perl-lsp-completion-item = { workspace = true }
+perl-path-security = { workspace = true }
+walkdir = "2.5.0"
+
+[lints]
+workspace = true
+
+[dev-dependencies]
+anyhow = "1.0.102"
