+++ crates/perl-lsp-file-path-completion/README.md
@@ -1,0 +1,3 @@
+# perl-lsp-file-path-completion
+
+Standalone SRP microcrate for secure, cancellation-aware file path completion items used by Perl LSP completion providers.
